requirements.txt
numpy==1.17.3
scipy==1.3.1
matplotlib==3.1.1
tqdm==4.38.0
scikit-learn==0.21.3
scikit-image==0.15.0
pillow==6.2.1
SimpleITK==1.2.3
tensorflow-gpu==2.0.0
keras==2.3.1
h5py==2.9.0